refactoring: deprecate some files
