+++ engine/a/engine_aaa/src/main/kotlin/net/crazyinvestor/engine_aaa/deprecated/service/TickerConsumerService.kt
@@ -1,0 +1,9 @@
+//package net.crazyinvestor.engine_aaa.service
+//
+//import net.crazyinvestor.engine_aaa.persistence.Ticker
+//import reactor.core.publisher.Mono
+//
+//// TODO deprecate 시키기
+//interface TickerConsumerService {
+//    fun createTicker(ticker: Ticker): Mono<Ticker>
+//}
