--- engine/a/engine_aaa/src/main/kotlin/net/crazyinvestor/engine_aaa/persistence/ExchangeRepository.kt
@@ -1,9 +1,0 @@
-package net.crazyinvestor.engine_aaa.persistence
-
-import org.springframework.data.cassandra.repository.ReactiveCassandraRepository
-import org.springframework.stereotype.Repository
-import java.util.UUID
-
-// TODO deprecate 시키기
-@Repository
-interface ExchangeRepository : ReactiveCassandraRepository<Exchange, UUID>
