--- engine/a/engine_aaa/src/main/kotlin/net/crazyinvestor/engine_aaa/service/TickerConsumerService.kt
@@ -1,9 +1,0 @@
-package net.crazyinvestor.engine_aaa.service
-
-import net.crazyinvestor.engine_aaa.persistence.Ticker
-import reactor.core.publisher.Mono
-
-// TODO deprecate 시키기
-interface TickerConsumerService {
-    fun createTicker(ticker: Ticker): Mono<Ticker>
-}
