+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/BithumbTickerEvent.kt
@@ -1,0 +1,6 @@
+package net.crazyinvestor.ticker_producer_engine.deprecated
+
+import net.crazyinvestor.ticker_producer_engine.dto.bithumb.response.BithumbWSTickerResponse
+import org.springframework.context.ApplicationEvent
+
+class BithumbTickerEvent(response: BithumbWSTickerResponse) : ApplicationEvent(response)
