+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/BithumbTransactionEvent.kt
@@ -1,0 +1,6 @@
+package net.crazyinvestor.ticker_producer_engine.deprecated
+
+import net.crazyinvestor.ticker_producer_engine.dto.bithumb.response.BithumbWSTransactionResponse
+import org.springframework.context.ApplicationEvent
+
+class BithumbTransactionEvent(response: BithumbWSTransactionResponse) : ApplicationEvent(response)
