+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/entity/Exchange.kt
@@ -1,0 +1,6 @@
+package net.crazyinvestor.ticker_producer_engine.deprecated.entity
+
+data class Exchange(
+    val id: Long?,
+    val name: String?
+)
