+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/entity/Quote.kt
@@ -1,4 +1,4 @@
-package net.crazyinvestor.ticker_producer_engine.entity
+package net.crazyinvestor.ticker_producer_engine.deprecated.entity
 
 data class Quote (
     var id: Long?,
