+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/entity/Ticker.kt
@@ -1,4 +1,4 @@
-package net.crazyinvestor.ticker_producer_engine.entity
+package net.crazyinvestor.ticker_producer_engine.deprecated.entity
 
 data class Ticker (
     val id: Long?,
