+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/repository/QuoteRepository.kt
@@ -1,7 +1,7 @@
 package net.crazyinvestor.ticker_producer_engine.deprecated.repository;
 
-import net.crazyinvestor.ticker_producer_engine.entity.Quote
-import net.crazyinvestor.ticker_producer_engine.entity.Symbol
+import net.crazyinvestor.ticker_producer_engine.deprecated.entity.Quote
+import net.crazyinvestor.ticker_producer_engine.deprecated.entity.Symbol
 import org.springframework.stereotype.Repository
 
 import java.util.Optional;
