+++ engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/deprecated/repository/SymbolRepository.kt
@@ -1,7 +1,7 @@
 package net.crazyinvestor.ticker_producer_engine.deprecated.repository
 
-import net.crazyinvestor.ticker_producer_engine.entity.Exchange
-import net.crazyinvestor.ticker_producer_engine.entity.Symbol
+import net.crazyinvestor.ticker_producer_engine.deprecated.entity.Exchange
+import net.crazyinvestor.ticker_producer_engine.deprecated.entity.Symbol
 import org.springframework.stereotype.Repository
 import java.util.*
