--- engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/entity/Exchange.kt
@@ -1,6 +1,0 @@
-package net.crazyinvestor.ticker_producer_engine.entity
-
-data class Exchange(
-    val id: Long?,
-    val name: String?
-)
