--- engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/event/BithumbOrderbookDepthEvent.kt
@@ -1,6 +1,0 @@
-package net.crazyinvestor.ticker_producer_engine.event
-
-import net.crazyinvestor.ticker_producer_engine.dto.response.BithumbWSOrderbookDepthResponse
-import org.springframework.context.ApplicationEvent
-
-class BithumbOrderbookDepthEvent(response: BithumbWSOrderbookDepthResponse) : ApplicationEvent(response)
