--- engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/event/BithumbTickerEvent.kt
@@ -1,6 +1,0 @@
-package net.crazyinvestor.ticker_producer_engine.event
-
-import net.crazyinvestor.ticker_producer_engine.dto.response.BithumbWSTickerResponse
-import org.springframework.context.ApplicationEvent
-
-class BithumbTickerEvent(response: BithumbWSTickerResponse) : ApplicationEvent(response)
