--- engine/a/ticker_producer_engine/src/main/kotlin/net/crazyinvestor/ticker_producer_engine/event/BithumbTransactionEvent.kt
@@ -1,6 +1,0 @@
-package net.crazyinvestor.ticker_producer_engine.event
-
-import net.crazyinvestor.ticker_producer_engine.dto.response.BithumbWSTransactionResponse
-import org.springframework.context.ApplicationEvent
-
-class BithumbTransactionEvent(response: BithumbWSTransactionResponse) : ApplicationEvent(response)
